Move common handler code to a separate file
